requirements.txt
nextcord
mysql-connector
yahooquery
matplotlib